Added a rotation when ship falls, added the maxEnergy and EnergyDecreaseRate to the config JSON

+++ Assets/COD/Scripts/Shared/ICollectable.cs
@@ -4,6 +4,6 @@ namespace COD.Shared
     {
         CollectableType Type { get; }
         int ScoreValue { get; }
-        int EnergyValue { get; set; }
+        float EnergyValue { get; set; }
     }
 }
